requirements.txt
# Arc Badlands CommPlex — Requirements v3.0
# Generated: 2026-04-19

# ── Core API ──────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0              # FastAPI test client + async HTTP

# ── Database ─────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
# Production (uncomment when moving to Cloud SQL):
# asyncpg>=0.29.0
# cloud-sql-python-connector[asyncpg]>=1.9.0

# ── Browser automation ────────────────────────────────────────────────────────
playwright>=1.42.0

# ── Google / Gmail / Workspace ────────────────────────────────────────────────
google-api-python-client>=2.120.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
gspread>=6.0.0             # Google Sheets live status

# ── GCP Services ─────────────────────────────────────────────────────────────
google-cloud-aiplatform>=1.47.0       # Vertex AI / Gemini Flash
google-cloud-bigquery>=3.17.0         # Campaign analytics
google-cloud-storage>=2.16.0          # GCS for PDFs + screenshots
google-cloud-secret-manager>=2.20.0   # Secret Manager
google-cloud-firestore>=2.16.0        # Hardware registry

# ── Telephony ────────────────────────────────────────────────────────────────
twilio>=8.13.0
requests>=2.31.0

# ── Data / Config ─────────────────────────────────────────────────────────────
pyyaml>=6.0
pandas>=2.2.0
openpyxl>=3.1.2

# ── Dev / Lint / Test ─────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.4.0
